Remove abstract audit log class
